refactor: Organização das pastas
