+++ lib/src/auth/authentication.dart
