+++ lib/src/pages/home/home_page.dart
@@ -1,12 +1,12 @@
 import 'package:firebase_auth/firebase_auth.dart'
     hide EmailAuthProvider, PhoneAuthProvider;
 import 'package:flutter/material.dart';
-import 'package:gtk_flutter/task_list.dart';
+import 'package:gtk_flutter/src/components/tasklist/task_list.dart';
 import 'package:provider/provider.dart';
 
-import 'app_state.dart';
-import 'src/authentication.dart';
-import 'src/widgets.dart';
+import '../../api/app_state.dart';
+import '../../auth/authentication.dart';
+import '../../auth/widgets.dart';
 
 class HomePage extends StatelessWidget {
   const HomePage({super.key});
@@ -39,7 +39,10 @@ class HomePage extends StatelessWidget {
               crossAxisAlignment: CrossAxisAlignment.start,
               children: [
                 if (appState.loggedIn) ...[
-                  TaskList(addTask: (description, title) => appState.addToTaskBoard(title, description), tasks: appState.taskListArray)
+                  TaskList(
+                      addTask: (description, title) =>
+                          appState.addToTaskBoard(title, description),
+                      tasks: appState.taskListArray)
                 ],
               ],
             ),
